# api-documentation.yaml
openapi: 3.0.0
info:
  title: PROgress Tracker API
  description: Breast Cancer Patient Survey Analytics Platform API
  version: 1.0.0
  contact:
    name: Support Team
    email: support@progresstracker.ca
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.progresstracker.ca
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          default: false
        message:
          type: string
        errors:
          type: array
          items:
            type: object
    Success:
      type: object
      properties:
        success:
          type: boolean
          default: true
        message:
          type: string
        data:
          type: object
    User:
      type: object
      properties:
        userId:
          type: string
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
        lastLogin:
          type: string
          format: date-time
...
    AnalyticsData:
      type: object
      properties:
        totalPatients:
          type: integer
        totalEvents:
          type: integer
        eventDistribution:
          type: array
          items:
            type: object
            properties:
              eventName:
                type: string
              count:
                type: integer
        completionRates:
          type: array
          items:
            type: object
paths:
...
  /api/auth/send-otp:
    post:
      tags:
        - Authentication
      summary: Send OTP to email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
      responses:
        200:
          description: OTP sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        400:
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/verify-otp:
    post:
      tags:
        - Authentication
      summary: Verify OTP and login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - otp
              properties:
                email:
                  type: string
                  format: email
                otp:
                  type: string
                  pattern: '^[0-9]{6}$'
      responses:
        200:
          description: Authentication successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        401:
          description: Invalid OTP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      security:
        - bearerAuth: []
      responses:
        200:
          description: Logged out successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
  /api/auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        200:
          description: User profile retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  user:
                    $ref: '#/components/schemas/User'

  # Upload Endpoints
...
  /api/analytics/dashboard:
    get:
      tags:
        - Analytics
      summary: Get dashboard analytics
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: timeRange
          schema:
            type: string
            enum: [day, week, month, year]
          description: Time range for analytics
      responses:
        200:
          description: Analytics data retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/AnalyticsData'
                  insights:
                    type: object
  /api/analytics/trends:
    get:
      tags:
        - Analytics
      summary: Get trend analysis
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: startDate
          schema:
            type: string
            format: date
        - in: query
          name: endDate
          schema:
            type: string
            format: date
        - in: query
          name: metric
          schema:
            type: string
      responses:
        200:
          description: Trends data retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
  /api/analytics/funnel:
    get:
      tags:
        - Analytics
      summary: Get funnel analysis
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: events
          required: true
          schema:
            type: string
          description: Comma-separated list of events
      responses:
        200:
          description: Funnel data retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
  /api/analytics/cohorts:
    get:
      tags:
        - Analytics
      summary: Get cohort analysis
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: cohortBy
          schema:
            type: string
            enum: [week, month]
        - in: query
          name: metric
          schema:
            type: string
      responses:
        200:
          description: Cohort data retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'

  # Patients Endpoints
...